Enable popovers and links on front page
